requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
filelock==3.0.12
fixture==1.5.11
idna==2.10
iniconfig==1.0.1
model-bakery==1.2.1
more-itertools==8.5.0
numpy==1.19.2
packaging==20.4
pandas==1.1.2
Pillow==8.0.1
pipenv==2020.11.15
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
sqlparse==0.4.1
toml==0.10.1
tqdm==4.47.0
urllib3==1.25.9
virtualenv==20.2.1
virtualenv-clone==0.5.4
vk-api==11.9.0
